@@ -9,26 +9,31 @@
     "start": "live-server --port=8081",
     "test": "jest"
   },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
+  "jest": {
+    "setupFilesAfterEnv": [
+      "./setup-tests.js"
+    ]
+  },
+  "author": "Jack Franklin",
+  "license": "MIT",
   "dependencies": {
-    "react": "15.3.2",
-    "react-dom": "15.3.2",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
     "shortid": "2.2.6"
   },
   "devDependencies": {
-    "babel-core": "6.14.0",
-    "babel-jest": "16.0.0",
-    "babel-loader": "6.2.5",
-    "babel-polyfill": "6.16.0",
-    "babel-preset-es2015": "6.16.0",
-    "babel-preset-react": "6.16.0",
-    "enzyme": "2.4.1",
-    "jest": "16.0.0",
+    "@babel/core": "^7.12.3",
+    "@babel/preset-env": "^7.12.1",
+    "@babel/preset-react": "^7.12.1",
+    "babel-jest": "^26.6.0",
+    "babel-loader": "^8.1.0",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.5",
+    "jest": "^26.6.0",
     "live-server": "1.1.0",
-    "react-addons-test-utils": "15.3.2",
+    "prettier": "^2.1.2",
     "react-test-renderer": "15.3.2",
-    "webpack": "1.13.2"
+    "webpack": "^5.1.3",
+    "webpack-cli": "^4.1.0"
   }
 }
